Moved Messages into new Django app
