+++ backend/Profile/models.py
@@ -1,7 +1,5 @@
 from django.db import models
 from django.contrib.auth.models import User
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from multiselectfield import MultiSelectField
