+++ backend/chat/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
 
-from .models import Chat, Message
+from .models import Chat
 
 admin.site.register(Chat)
-admin.site.register(Message)
