+++ backend/events/urls.py
@@ -5,6 +5,7 @@ router = routers.DefaultRouter()
 router.register('', EventViewSet, 'events')
 
 urlpatterns = [
-            path('^admin/', admin.site.urls),
-            path('^api/', include(router.urls))
+    path('^admin/', admin.site.urls),
+    path('^api/', include(router.urls))
 ]
+-
